renaming database modules
