+++ lib/ansible/modules/cloud/azure/azure_rm_postgresqlfirewallrule_info.py
@@ -1,7 +1,6 @@
 #!/usr/bin/python
 #
-# Copyright (c) 2017 Zim Kalinowski, <zikalino@microsoft.com>
-# Copyright (c) 2019 Matti Ranta, (@techknowlogick)
+# Copyright (c) 2018 Zim Kalinowski, <zikalino@microsoft.com>
 #
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
@@ -16,16 +15,16 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 
 DOCUMENTATION = '''
 ---
-module: azure_rm_mariadbdatabase_facts
+module: azure_rm_postgresqlfirewallrule_info
 version_added: "2.8"
-short_description: Get Azure MariaDB Database facts.
+short_description: Get Azure PostgreSQL Firewall Rule facts.
 description:
-    - Get facts of MariaDB Database.
+    - Get facts of Azure PostgreSQL Firewall Rule.
 
 options:
     resource_group:
         description:
-            - The name of the resource group that contains the resource. You can obtain this value from the Azure Resource Manager API or the portal.
+            - The name of the resource group.
         required: True
     server_name:
         description:
@@ -33,33 +32,32 @@ options:
         required: True
     name:
         description:
-            - The name of the database.
+            - The name of the server firewall rule.
 
 extends_documentation_fragment:
     - azure
 
 author:
     - "Zim Kalinowski (@zikalino)"
-    - "Matti Ranta (@techknowlogick)"
 
 '''
 
 EXAMPLES = '''
-  - name: Get instance of MariaDB Database
-    azure_rm_mariadbdatabase_facts:
+  - name: Get instance of PostgreSQL Firewall Rule
+    azure_rm_postgresqlfirewallrule_info:
       resource_group: myResourceGroup
       server_name: server_name
-      name: database_name
+      name: firewall_rule_name
 
-  - name: List instances of MariaDB Database
-    azure_rm_mariadbdatabase_facts:
+  - name: List instances of PostgreSQL Firewall Rule
+    azure_rm_postgresqlfirewallrule_info:
       resource_group: myResourceGroup
       server_name: server_name
 '''
 
 RETURN = '''
-databases:
-    description: A list of dictionaries containing facts for MariaDB Databases.
+rules:
+    description: A list of dictionaries containing facts for PostgreSQL Firewall Rule.
     returned: always
     type: complex
     contains:
@@ -68,17 +66,11 @@ databases:
                 - Resource ID
             returned: always
             type: str
-            sample: "/subscriptions/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/resourceGroups/myResourceGroup/providers/Microsoft.DBforMariaDB/servers/testser
-                    ver/databases/db1"
-        resource_group:
-            description:
-                - Resource group name.
-            returned: always
-            type: str
-            sample: testrg
+            sample: "/subscriptions/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/resourceGroups/TestGroup/providers/Microsoft.DBforPostgreSQL/servers/testserver/fire
+                    wallRules/rule1"
         server_name:
             description:
-                - Server name.
+                - The name of the server.
             returned: always
             type: str
             sample: testserver
@@ -87,33 +79,33 @@ databases:
                 - Resource name.
             returned: always
             type: str
-            sample: db1
-        charset:
+            sample: rule1
+        start_ip_address:
             description:
-                - The charset of the database.
+                - The start IP address of the PostgreSQL firewall rule.
             returned: always
             type: str
-            sample: UTF8
-        collation:
+            sample: 10.0.0.16
+        end_ip_address:
             description:
-                - The collation of the database.
+                - The end IP address of the PostgreSQL firewall rule.
             returned: always
             type: str
-            sample: English_United States.1252
+            sample: 10.0.0.18
 '''
 
 from ansible.module_utils.azure_rm_common import AzureRMModuleBase
 
 try:
     from msrestazure.azure_exceptions import CloudError
-    from azure.mgmt.rdbms.mariadb import MariaDBManagementClient
+    from azure.mgmt.rdbms.postgresql import PostgreSQLManagementClient
     from msrest.serialization import Model
 except ImportError:
     # This is handled in azure_rm_common
     pass
 
 
-class AzureRMMariaDbDatabaseFacts(AzureRMModuleBase):
+class AzureRMPostgreSQLFirewallRulesFacts(AzureRMModuleBase):
     def __init__(self):
         # define user inputs into argument
         self.module_arg_spec = dict(
@@ -133,34 +125,34 @@ class AzureRMMariaDbDatabaseFacts(AzureRMModuleBase):
         self.results = dict(
             changed=False
         )
+        self.mgmt_client = None
         self.resource_group = None
         self.server_name = None
         self.name = None
-        super(AzureRMMariaDbDatabaseFacts, self).__init__(self.module_arg_spec, supports_tags=False)
+        super(AzureRMPostgreSQLFirewallRulesFacts, self).__init__(self.module_arg_spec, supports_tags=False)
 
     def exec_module(self, **kwargs):
         for key in self.module_arg_spec:
             setattr(self, key, kwargs[key])
+        self.mgmt_client = self.get_mgmt_svc_client(PostgreSQLManagementClient,
+                                                    base_url=self._cloud_environment.endpoints.resource_manager)
 
-        if (self.resource_group is not None and
-                self.server_name is not None and
-                self.name is not None):
-            self.results['databases'] = self.get()
-        elif (self.resource_group is not None and
-              self.server_name is not None):
-            self.results['databases'] = self.list_by_server()
+        if (self.name is not None):
+            self.results['rules'] = self.get()
+        else:
+            self.results['rules'] = self.list_by_server()
         return self.results
 
     def get(self):
         response = None
         results = []
         try:
-            response = self.mariadb_client.databases.get(resource_group_name=self.resource_group,
-                                                         server_name=self.server_name,
-                                                         database_name=self.name)
+            response = self.mgmt_client.firewall_rules.get(resource_group_name=self.resource_group,
+                                                           server_name=self.server_name,
+                                                           firewall_rule_name=self.name)
             self.log("Response : {0}".format(response))
         except CloudError as e:
-            self.log('Could not get facts for Databases.')
+            self.log('Could not get facts for FirewallRules.')
 
         if response is not None:
             results.append(self.format_item(response))
@@ -171,11 +163,11 @@ class AzureRMMariaDbDatabaseFacts(AzureRMModuleBase):
         response = None
         results = []
         try:
-            response = self.mariadb_client.databases.list_by_server(resource_group_name=self.resource_group,
-                                                                    server_name=self.server_name)
+            response = self.mgmt_client.firewall_rules.list_by_server(resource_group_name=self.resource_group,
+                                                                      server_name=self.server_name)
             self.log("Response : {0}".format(response))
         except CloudError as e:
-            self.fail("Error listing for server {0} - {1}".format(self.server_name, str(e)))
+            self.log('Could not get facts for FirewallRules.')
 
         if response is not None:
             for item in response:
@@ -187,16 +179,17 @@ class AzureRMMariaDbDatabaseFacts(AzureRMModuleBase):
         d = item.as_dict()
         d = {
             'resource_group': self.resource_group,
+            'id': d['id'],
             'server_name': self.server_name,
             'name': d['name'],
-            'charset': d['charset'],
-            'collation': d['collation']
+            'start_ip_address': d['start_ip_address'],
+            'end_ip_address': d['end_ip_address']
         }
         return d
 
 
 def main():
-    AzureRMMariaDbDatabaseFacts()
+    AzureRMPostgreSQLFirewallRulesFacts()
 
 
 if __name__ == '__main__':
